fix translation of terminators in MSVC cleanup blocks
I have taken a fairly brute force approach here: for each successor, I
generate a basic block that contains a `cleanupret` to that block. It
is ugly but it works.

It might have been possible to play with phis, but there are rumours that
there must be a 1-1 mapping between cleanup pad entries and exits, which
is not the case in MIR.